Add Speed Detections
This speed detection will have an expected travel distance, and if it goes above this, it will mark the player.
Taking into account lag, if the player's ping is high, and they seem to teleport around, the lastTime will take in account for that. Every tick that the player does not move, it will increase the expected by 0.8..., multiplying the amount of ticks not moving by the expected. This will also be handy for server lag and TPS drops, since micro time does not depend on the server.